Добавить тест testFindUserById и пофиксить баги :)

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -25,7 +25,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User findById(@PathVariable Long id) {
+    public User findById(@PathVariable int id) {
         return userService.findById(id);
     }
 
@@ -40,22 +40,22 @@ public class UserController {
     }
 
     @GetMapping("/{id}/friends")
-    public Collection<User> getFriends(@PathVariable Long id) {
+    public Collection<User> getFriends(@PathVariable int id) {
         return userService.getFriends(id);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public Collection<User> findSharedFriends(@PathVariable Long id, @PathVariable Long otherId) {
+    public Collection<User> findSharedFriends(@PathVariable int id, @PathVariable int otherId) {
         return userService.findSharedFriend(id, otherId);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    public Collection<Long> addFriend(@PathVariable Long id, @PathVariable Long friendId) {
+    public Collection<Integer> addFriend(@PathVariable int id, @PathVariable int friendId) {
         return userService.addUserFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    public Collection<Long> deleteFriend(@PathVariable Long id, @PathVariable Long friendId) {
+    public Collection<Integer> deleteFriend(@PathVariable int id, @PathVariable int friendId) {
         return userService.deleteUserFriend(id, friendId);
     }
 }
