+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -10,19 +10,19 @@ import java.util.Set;
 
 @Data
 public class User {
-    private Long id;
+    private Integer id;
     private String email;
     private String login;
     private String name;
     private LocalDate birthday;
-    private Map<Long, Boolean> friends = new HashMap<>();
+    private Map<Integer, Boolean> friends = new HashMap<>();
 
-    public Set<Long> addFriend(User user) {
+    public Set<Integer> addFriend(User user) {
         friends.put(user.getId(), false);
         return friends.keySet();
     }
 
-    public Set<Long> deleteFriend(User user) {
+    public Set<Integer> deleteFriend(User user) {
         friends.remove(user.getId());
         return friends.keySet();
     }
